Brazilian Online Marketing Platform Olist
Looking into the Brazilian online market platform where people place orders from around the world. Much has to be done on the dataset as it include 6 different tables with much to clean. But in the long run, I was able to generate an insight as to the progress of the site overtime to help make informed decision as to their payment, deliveries, products and vendors. 
Tool PowerBI
Project Duration 24hrs
Time 2022