The Shark Attack
The myth about Shark attacking human can be understood with this report. The Data very much unorganized as it must have been collated thoroughly. Much needed to be done as to the kind of shark in a region, the attack time, the body parts, death rate, death by gender and so on. Aside from being a dirty data, it has missing values and so.
https://public.tableau.com/app/profile/ebenezer.abiola.ogunyinka/viz/SharkAttack_16722700173550/AttackThirdOverview?publish=yes
Tool: PowwerBI and Tableau
Duration: 4 days
Time: 2021 and re-assigned in 2022