2015 Delayed Flights in US
Aviation industry in the United state wanted to know why so many flights were delayed in 2015.  The dataset presented has so much to make this possible but very clumsy with voluminous rows and multiple tables. Since the focus is the delayed flights, therefore, cleaning was done, filtering, modeling the tables and I have to separate the delayed flights from the on-time flight, as well as the early departures, early arrival and the like. The Dashboard report gives an insight into the dataset for effective conclusion as to Boston Airport Delayed history. 
Tool: PowerBI
Duration: 92 Hrs
Time 2022/2023