World Energy Consumption
This dataset is just numbers and nothing else. The Date column needed some modifications for us to see the flow of energy consumption over the years. With much effort and carful study of the Dataset, much was achieved to give an understanding into which kind of energy consumed per region and so on. 
https://public.tableau.com/app/profile/ebenezer.abiola.ogunyinka/viz/WorldEnergyConsumption_16726074288600/ExecutiveOverview?publish=yes
Tool: Tableau
Duration: 2 Days 
Time: 2022 and 2023